Extract modal focus trap hook
